// src/types/EffectDefinition.ts
export enum EffectDefinition {
  BuildupTeleportToOverworld = 42574,
  DamageIndirectEffect = 654,
  DamageProtectionIndirectEffect = 64858,
  Feed = 3280,
  Heal = 48456,
  Inate = 22728,
  LuminosityIndirectEffect = 65294,
  Nourish = 20046,
  PoisonoverTime = 29250,
  RevealEffect = 62720,
  SpeedIndirectEffect = 6764,
  TeleportToOverworld = 27214
}