// src/types/LiquidVisualChunkDefinition.ts
export enum LiquidVisualChunkDefinition {
  AppleBurnt = 56934,
  AppleCooked = 56516,
  AppleRipe = 56016,
  AppleUnripe = 55460,
  BabuBurnt = 12722,
  BabuCooked = 17228,
  BabuRipe = 21734,
  BlueberryBurnt = 13156,
  BlueberryCooked = 9572,
  BlueberryRipe = 5988,
  BlueberryUnripe = 2404,
  CarrotBurnt = 1502,
  CarrotCooked = 50674,
  CarrotRipe = 50666,
  CarrotUnripe = 1510,
  EggplantBurnt = 50610,
  EggplantCooked = 50602,
  EggplantRipe = 50594,
  EggplantUnripe = 50584,
  GarlicBurnt = 50570,
  GarlicCooked = 50562,
  GarlicRipe = 50554,
  GarlicUnripe = 50544,
  MushroomBrownBurnt = 39088,
  MushroomBrownCooked = 20040,
  MushroomBrownRipe = 21422,
  MushroomCaveLargeBurnt = 58130,
  MushroomCaveLargeCooked = 52698,
  MushroomCaveLargeRipe = 47266,
  MushroomCaveSmallBurnt = 38392,
  MushroomCaveSmallCooked = 32334,
  MushroomCaveSmallRipe = 26276,
  MushroomGlowingBurnt = 9532,
  MushroomGlowingCooked = 4142,
  MushroomGlowingRipe = 64288,
  MushroomRedBurnt = 32604,
  MushroomRedCooked = 27048,
  MushroomRedRipe = 21492,
  OnionBurnt = 50530,
  OnionCooked = 50522,
  OnionRipe = 50514,
  OnionUnripe = 50504,
  PotatoBurnt = 13336,
  PotatoCooked = 13360,
  PotatoRipe = 13352,
  PotatoUnripe = 13344,
  PumpkinBurnt = 30124,
  PumpkinCooked = 29434,
  PumpkinRipe = 28744,
  PumpkinUnripe = 28054,
  Salt = 13596,
  SpriggullDrumstickBurnt = 32864,
  SpriggullDrumstickCooked = 27542,
  SpriggullDrumstickRipe = 22220,
  TomatoBurnt = 50490,
  TomatoCooked = 50482,
  TomatoRipe = 50474,
  TomatoUnripe = 50464
}