// src/types/LiquidVisualData.ts
export enum LiquidVisualData {
  BlueMixStewBurnt = 33594,
  BlueMixStewCooked = 13166,
  BlueMixStewRaw = 51552,
  BlueberryStewBurnt = 22532,
  BlueberryStewCooked = 22516,
  BlueberryStewRaw = 41466,
  DefaultStewBurnt = 22476,
  DefaultStewCooked = 22452,
  DefaultStewRaw = 21644,
  EggplantPotatoRedStewBurnt = 13690,
  EggplantPotatoRedStewCooked = 36046,
  EggplantPotatoRedStewRaw = 47072,
  GlowingMushroomStewBurnt = 21060,
  GlowingMushroomStewCooked = 21046,
  GlowingMushroomStewRaw = 20952,
  MushroomSpriggullStewBurnt = 22598,
  MushroomSpriggullStewCooked = 22590,
  MushroomSpriggullStewRaw = 22572,
  MushroomStewBurnt = 21600,
  MushroomStewCooked = 21586,
  MushroomStewRaw = 21578,
  NonRecipeStewBurnt = 21202,
  NonRecipeStewCooked = 21182,
  NonRecipeStewRaw = 21158,
  PotatoPureeStewCooked = 40124,
  PotatoPureeStewRaw = 6764,
  PoultryOnionMushroomStewCooked = 15654,
  PoultryOnionMushroomStewRaw = 45270,
  PoultryPotatoStewCooked = 28868,
  PoultryPotatoStewRaw = 60754,
  PumpkinStewBurnt = 62608,
  PumpkinStewCooked = 62618,
  PumpkinStewRaw = 62628,
  RedMushroomStewBurnt = 21248,
  RedMushroomStewCooked = 21230,
  RedMushroomStewRaw = 28546,
  RedSauceStewCooked = 45938,
  RedSauceStewRaw = 17112,
  SpriggullStewBurnt = 41908,
  SpriggullStewCooked = 41578,
  SpriggullStewRaw = 41624,
  Teleport = 49754,
  VegetableRaguStewCooked = 38666,
  VegetableRaguStewRaw = 13246,
  VisionStewCooked = 55814,
  VisionStewRaw = 24160,
  Water = 7838
}